Index Qoder CLI sessions in the 2.0 app as well
The 2.0 Qoder loader scanned only `cache/projects/<slug>/conversation-history`
(Qoder IDE), so transcripts the Qoder CLI writes under `projects/<slug>` were
never indexed. Session sources have to behave the same in both apps, so mirror
the 1.0 fix here.

Reusing the Claude-Code parser required breaking a cycle: `session-loader.ts`
already imports from `session-loaders/alternative-sources.ts`, so the Claude
row parsing moves into `session-loaders/claude-cli.ts`, which both can import.
