+++ apps/main-2.0/src/core/session-loader-qoder.test.ts
@@ -1,0 +1,129 @@
+import { afterEach, describe, expect, it } from "vitest";
+import * as fs from "node:fs";
+import * as os from "node:os";
+import * as path from "node:path";
+
+import { loadQoderSessions } from "./session-loader";
+
+const temporaryRoots: string[] = [];
+
+function temporaryRoot(label: string): string {
+  const root = fs.mkdtempSync(path.join(os.tmpdir(), `agentrecall-qoder-v2-${label}-`));
+  temporaryRoots.push(root);
+  return root;
+}
+
+function writeJsonl(filePath: string, rows: unknown[]): void {
+  fs.mkdirSync(path.dirname(filePath), { recursive: true });
+  fs.writeFileSync(filePath, rows.map((row) => JSON.stringify(row)).join("\n"));
+}
+
+afterEach(() => {
+  for (const root of temporaryRoots.splice(0)) {
+    fs.rmSync(root, { recursive: true, force: true });
+  }
+});
+
+describe("Qoder session loading", () => {
+  it("loads Qoder CLI transcripts stored directly under projects/<slug>", () => {
+    const root = temporaryRoot("cli-flat");
+    writeJsonl(path.join(root, "projects", "-Users-me-demo-app", "5a5f525e-99bc-4c95-9f03-de30ef8c9a32.jsonl"), [
+      { type: "workspace-directories", sessionId: "5a5f525e", directories: ["/Users/me/demo-app"] },
+      {
+        type: "user",
+        uuid: "u1",
+        parentUuid: null,
+        cwd: "/Users/me/demo-app",
+        gitBranch: "main",
+        timestamp: "2026-08-21T10:00:00.000Z",
+        message: { role: "user", content: [{ type: "text", text: "Fix the login bug" }] },
+      },
+      {
+        type: "assistant",
+        uuid: "a1",
+        parentUuid: "u1",
+        cwd: "/Users/me/demo-app",
+        timestamp: "2026-08-21T10:00:05.000Z",
+        message: { role: "assistant", content: [{ type: "text", text: "I will check the auth module." }] },
+      },
+    ]);
+
+    const loaded = loadQoderSessions(root);
+
+    expect(loaded).toHaveLength(1);
+    expect(loaded[0].session).toMatchObject({
+      sessionKey: "qoder:5a5f525e-99bc-4c95-9f03-de30ef8c9a32",
+      source: "qoder",
+      projectPath: "/Users/me/demo-app",
+      gitBranch: "main",
+      firstQuestion: "Fix the login bug",
+    });
+    expect(loaded[0].messages.map((message) => message.role)).toEqual(["user", "assistant"]);
+  });
+
+  it("loads Qoder CLI transcripts from the nested transcript directory and prefers ai-title", () => {
+    const root = temporaryRoot("cli-nested");
+    writeJsonl(path.join(root, "projects", "-Users-me-demo-app", "transcript", "428f5d29.jsonl"), [
+      { type: "ai-title", sessionId: "428f5d29", aiTitle: "Refactor the auth module" },
+      {
+        type: "user",
+        uuid: "u1",
+        parentUuid: null,
+        cwd: "/Users/me/demo-app",
+        message: { role: "user", content: [{ type: "text", text: "Please refactor auth" }] },
+      },
+    ]);
+
+    const loaded = loadQoderSessions(root);
+
+    expect(loaded).toHaveLength(1);
+    expect(loaded[0].session.rawId).toBe("428f5d29");
+    expect(loaded[0].session.originalTitle).toBe("Refactor the auth module");
+  });
+
+  it("skips Qoder CLI tool-execution traces that contain no readable messages", () => {
+    const root = temporaryRoot("cli-trace");
+    writeJsonl(path.join(root, "projects", "-Users-me-demo-app", "transcript", "trace-only.jsonl"), [
+      {
+        type: "user",
+        uuid: "u1",
+        parentUuid: null,
+        cwd: "/Users/me/demo-app",
+        message: { role: "user", content: [{ type: "tool_result", tool_use_id: "t1", content: "ok" }] },
+      },
+      { type: "progress", uuid: "p1", parentUuid: "u1" },
+      {
+        type: "assistant",
+        uuid: "a1",
+        parentUuid: "u1",
+        cwd: "/Users/me/demo-app",
+        message: { role: "assistant", content: [{ type: "tool_use", id: "t2", name: "Read", input: {} }] },
+      },
+    ]);
+
+    expect(loadQoderSessions(root)).toEqual([]);
+  });
+
+  it("indexes Qoder IDE and CLI layouts together and skips unrecognised JSONL", () => {
+    const root = temporaryRoot("both");
+    writeJsonl(path.join(root, "cache", "projects", "demo-app-1a2b3c4d", "conversation-history", "task-ide", "task-ide.jsonl"), [
+      { role: "user", message: { content: [{ type: "text", text: "IDE question" }] } },
+    ]);
+    writeJsonl(path.join(root, "projects", "-Users-me-demo-app", "cli-session.jsonl"), [
+      {
+        type: "user",
+        uuid: "u1",
+        parentUuid: null,
+        cwd: "/Users/me/demo-app",
+        message: { role: "user", content: [{ type: "text", text: "CLI question" }] },
+      },
+    ]);
+    writeJsonl(path.join(root, "projects", "-Users-me-demo-app", "compaction", "not-a-transcript.jsonl"), [
+      { kind: "summary", tokens: 42 },
+    ]);
+
+    const loaded = loadQoderSessions(root);
+
+    expect(loaded.map((entry) => entry.session.firstQuestion).sort()).toEqual(["CLI question", "IDE question"]);
+  });
+});
