+++ apps/main-2.0/src/core/session-loader.ts
@@ -36,6 +36,15 @@ export * from "./session-loaders/alternative-sources";
 import { loadWorkBuddySessionsIterator } from "./session-loaders/workbuddy";
 export * from "./session-loaders/workbuddy";
 import {
+  claudeVisibleConversationRows,
+  extractClaudeTokenEvents,
+  extractClaudeTraceEvents,
+  firstAiTitle,
+  firstClaudeGitBranch,
+  loadClaudeCliSessionRows,
+} from "./session-loaders/claude-cli";
+export * from "./session-loaders/claude-cli";
+import {
   createIndexedSession,
   createTokenUsage,
   dedupeTraceEvents,
@@ -316,81 +325,6 @@ function codexVisibleConversationRows(rows: unknown[]): unknown[] {
   }
 
   return [...preamble, ...turns.flatMap((turn) => turn.rows)];
-}
-
-function claudeVisibleConversationRows(rows: unknown[]): unknown[] {
-  const conversationRows = rows.filter((row) => isRecord(row) && (row.type === "user" || row.type === "assistant"));
-  if (conversationRows.length === 0) return rows;
-
-  const nodes = new Map<string, Record<string, unknown>>();
-  for (const row of conversationRows) {
-    if (!isRecord(row)) return rows;
-    const uuid = stringField(row, "uuid");
-    if (!uuid || nodes.has(uuid)) return rows;
-    nodes.set(uuid, row);
-  }
-
-  const visibleUuids = new Set<string>();
-  let current = conversationRows.at(-1) as Record<string, unknown>;
-  while (current) {
-    const uuid = stringField(current, "uuid");
-    if (!uuid || visibleUuids.has(uuid)) return rows;
-    visibleUuids.add(uuid);
-    const parentUuid = unknownField(current, "parentUuid");
-    if (parentUuid === null || parentUuid === undefined || parentUuid === "") break;
-    if (typeof parentUuid !== "string") return rows;
-    const parent = nodes.get(parentUuid);
-    if (!parent) return rows;
-    current = parent;
-  }
-
-  return rows.filter((row) => {
-    if (!isRecord(row) || (row.type !== "user" && row.type !== "assistant")) return true;
-    return visibleUuids.has(stringField(row, "uuid"));
-  });
-}
-
-function extractClaudeTraceEvents(rows: unknown[]): TraceEventDraft[] {
-  const events: TraceEventDraft[] = [];
-
-  for (const row of rows) {
-    if (!isRecord(row) || (row.type !== "user" && row.type !== "assistant")) continue;
-    const message = objectField(row, "message");
-    const blocks = unknownField(message, "content");
-    if (!Array.isArray(blocks)) continue;
-
-    for (const block of blocks) {
-      if (!isRecord(block)) continue;
-      if (block.type === "tool_use") {
-        const input = unknownField(block, "input");
-        const name = stringField(block, "name") || "tool";
-        const summary = firstStringField(input, ["command", "cmd", "file_path", "path", "query", "url"]);
-        events.push({
-          kind: "tool_call",
-          source: "claude",
-          title: titleWithSummary(name, summary),
-          detail: stringifyDetail(input),
-          timestamp: stringField(row, "timestamp"),
-          callId: stringField(block, "id") || null,
-          eventType: null,
-          status: "running",
-        });
-      } else if (block.type === "tool_result") {
-        events.push({
-          kind: "tool_result",
-          source: "claude",
-          title: "tool result",
-          detail: stringifyDetail(unknownField(block, "content")),
-          timestamp: stringField(row, "timestamp"),
-          callId: stringField(block, "tool_use_id") || null,
-          eventType: null,
-          status: unknownField(block, "is_error") === true ? "failed" : "completed",
-        });
-      }
-    }
-  }
-
-  return events;
 }
 
 function extractCodexResponseTrace(
@@ -896,40 +830,6 @@ function extractCodexTokenEvents(rows: readonly CodexTokenRow[]): TokenUsageEven
   return cumulativeEntries.size > 0 ? [...cumulativeEntries.values()] : [...entries.values()];
 }
 
-function extractClaudeTokenEvents(rows: unknown[]): TokenUsageEvent[] {
-  const entries = new Map<string, TokenUsageEvent>();
-
-  rows.forEach((row, index) => {
-    if (!isRecord(row) || row.type !== "assistant") return;
-    const message = objectField(row, "message");
-    const usage = objectField(message, "usage");
-    if (!usage) return;
-
-    const cached =
-      numberField(usage, "cache_read_input_tokens") +
-      numberField(usage, "cached_input_tokens");
-    const cacheCreation = numberField(usage, "cache_creation_input_tokens");
-    const entry = createTokenUsage(
-      numberField(usage, "input_tokens"),
-      numberField(usage, "output_tokens"),
-      cached,
-      numberField(usage, "reasoning_output_tokens"),
-      cacheCreation,
-    );
-    const key = stringField(message, "id") || stringField(row, "uuid") || `${index}:${JSON.stringify(usage)}`;
-    putTokenEvent(
-      entries,
-      {
-        ...entry,
-        timestamp: parseTimestampMs(row.timestamp),
-        dedupeKey: key.startsWith("claude-code:") ? key : `claude-code:${key}`,
-      },
-    );
-  });
-
-  return [...entries.values()];
-}
-
 function extractCodeBuddyTokenEvents(rows: unknown[]): TokenUsageEvent[] {
   const entries = new Map<string, TokenUsageEvent>();
 
@@ -1007,15 +907,6 @@ function firstDetailNumber(value: unknown, key: string): number {
   return numberField(value, key);
 }
 
-function firstClaudeGitBranch(rows: unknown[]): string | null {
-  for (const row of rows) {
-    if (!row || typeof row !== "object" || !("gitBranch" in row)) continue;
-    const branch = (row as ClaudeConversationLine).gitBranch?.trim();
-    if (branch) return branch;
-  }
-  return null;
-}
-
 // CodeBuddy rows do not currently embed gitBranch, so derive it from the
 // session working directory when that directory belongs to a Git repository.
 function readGitBranchFromCwd(cwd: string): string | null {
@@ -1077,17 +968,6 @@ function firstCodeBuddySessionMeta(rows: unknown[], fallbackRawId: string): { ra
   }
 
   return { rawId, projectPath, timestamp };
-}
-
-// Claude and CodeBuddy write their AI-generated session title as a dedicated
-// `ai-title` row. The row is metadata and is not exposed as a visible message.
-function firstAiTitle(rows: unknown[]): string {
-  for (const row of rows) {
-    if (!isRecord(row) || row.type !== "ai-title") continue;
-    const title = stringField(row, "aiTitle").trim();
-    if (title) return title;
-  }
-  return "";
 }
 
 export function loadCodexSessionRows(
@@ -1581,66 +1461,6 @@ export async function* loadCodexSessionsAsyncIterator(
 
 function encodeClaudeProjectDir(cwd: string): string {
   return cwd.replace(/[^a-zA-Z0-9-]/g, "-");
-}
-
-export function loadClaudeCliSessionRows(
-  filePath: string,
-  rows: unknown[],
-  options: {
-    rawId?: string;
-    cwd?: string;
-    startedAt?: number;
-    source?: SessionSource;
-    stepcodeAgent?: "claude" | "codex";
-    stat?: VirtualSessionFileStat;
-    isSubagent?: boolean;
-    parentSessionId?: string | null;
-    includeTraceEvents?: boolean;
-  } = {},
-): LoadedSession | null {
-  const rawId = options.rawId || path.basename(filePath, ".jsonl");
-  const visibleRows = claudeVisibleConversationRows(rows);
-  const messages = extractMessages(visibleRows, "claude");
-  const tokenEvents = extractClaudeTokenEvents(rows);
-  const traceEvents = options.includeTraceEvents === false ? [] : extractTraceEvents(visibleRows, "claude");
-  const tokenUsage = tokenUsageFromEvents(tokenEvents);
-  const question = firstQuestion(messages);
-  let customTitle = "";
-  for (let index = rows.length - 1; index >= 0; index -= 1) {
-    const row = rows[index];
-    if (!isRecord(row) || row.type !== "custom-title") continue;
-    customTitle = stringField(row, "customTitle").trim();
-    if (customTitle) break;
-  }
-  const aiTitle = firstAiTitle(rows);
-  const embeddedCwd = (rows.find((row) => row && typeof row === "object" && "cwd" in row) as ClaudeConversationLine | undefined)?.cwd;
-  const gitBranch = firstClaudeGitBranch(rows);
-  const source = options.source
-    ?? (options.stepcodeAgent === "claude" ? "stepcode-claude" : "claude-cli");
-  return {
-    session: createIndexedSession({
-      keyPrefix: source === "tclaude-cli"
-        ? "tclaude"
-        : source === "stepcode-claude"
-          ? "stepcode-claude"
-          : "claude",
-      rawId,
-      source,
-      projectPath: options.cwd || embeddedCwd || "",
-      filePath,
-      originalTitle: customTitle || aiTitle || cleanTitle(question) || "Untitled Session",
-      firstQuestion: cleanTitle(question),
-      timestamp: options.startedAt || 0,
-      gitBranch,
-      tokenUsage,
-      stat: options.stat,
-      isSubagent: options.isSubagent,
-      parentSessionId: options.parentSessionId,
-    }),
-    messages,
-    tokenEvents,
-    traceEvents,
-  };
 }
 
 export function loadClaudeCliSessions(claudeDir = path.join(os.homedir(), ".claude"), source: SessionSource = "claude-cli"): LoadedSession[] {
